# backend/openapi.files.gridfs.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: "CRUD-with-files-mongodb-gridfs"
servers:
  - url: "localhost:5000"
paths:
  /api/files/gridfs/images:
    get:
      summary: "Get all images in binary format from a bucket"
      operationId: getImages
      responses:
        "200":
          description: "All images are fetched successfully"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "500":
          description: "Unexpected error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: { code: 500, message: "Something unexpected happened" }
    post:
      summary: "Add a new image"
      operationId: postImage
      requestBody:
        description: "The whole image in binary"
        content:
          multipart/form-data:
            schema:
              properties:
                image:
                  type: string
                  format: binary
      responses:
        "201":
          description: "Image uploaded"
          content:
            application/json:
              schema:
                type: string
              example: "/uploads/images/4.jpg"
        "400":
          description: File is not in jpg or png format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                {
                  code: 400,
                  message: "The file you tried to upload is not in jpg or png format"
                }
        "500":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: { code: 500, message: "Something unexpected happened" }
  /api/files/gridfs/images/{id}:
    get:
      summary: "Get specific image in binary format from a bucket"
      operationId: getImage
      parameters:
        - name: id
          in: path
          description: ID of the image
          required: true
          schema:
            type: integer
            format: int32
          example: 1
      responses:
        "200":
          description: "Image is fetched successfully"
    delete:
      summary: "Delete an image from the server"
      operationId: deleteImage
      parameters:
        - name: id
          in: path
          description: ID of the image
          required: true
          schema:
            type: integer
            format: int32
          example: 1
      responses:
        "200":
          description: "Image successfully deleted"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              example: ["/uploads/images/2.png", "/uploads/images/3.gif"]
        "404":
          description: "Image not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                {
                  code: 404,
                  message: "The image you are trying to delete is not found"
                }
  /api/files/gridfs/audio:
    get:
      summary: "Get all audio files"
      operationId: getAudio
      responses:
        "200":
          description: "All audio files are fetched successfully"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              example:
                ["/uploads/audio/1", "/uploads/audio/2", "/uploads/audio/3"]
        "500":
          description: "Unexpected error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: { code: 500, message: "Something unexpected happened" }
    post:
      summary: "Add a new audio file"
      operationId: postAudio
      requestBody:
        description: "The whole audio in binary"
        content:
          multipart/form-data:
            schema:
              properties:
                audio:
                  type: string
                  format: binary
      responses:
        "201":
          description: "Audio uploaded"
          content:
            application/json:
              schema:
                type: string
              example: "/uploads/audio/4"
        "500":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: { code: 500, message: "Something unexpected happened" }
  /api/files/gridfs/audio/{id}:
    get:
      summary: "Get specific audio file in binary format from a bucket"
      operationId: getAudioFile
      parameters:
        - name: id
          in: path
          description: ID of the image
          required: true
          schema:
            type: integer
            format: int32
          example: 1
      responses:
        "200":
          description: "Audio is fetched successfully"
    delete:
      summary: "Delete an audio from the server"
      operationId: deleteAudio
      parameters:
        - name: id
          in: path
          description: ID of the audio
          required: true
          schema:
            type: integer
            format: int32
          example: 1
      responses:
        "200":
          description: "Audio successfully deleted"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              example: ["/uploads/audio/2", "/uploads/audio/3"]
        "404":
          description: "Audio not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                {
                  code: 404,
                  message: "The audio file you are trying to delete is not found"
                }
  /api/files/gridfs/video:
    get:
      summary: "Get all videos"
      operationId: getVideo
      responses:
        "200":
          description: "All video links are fetched successfully"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              example:
                ["/uploads/video/1", "/uploads/video/2", "/uploads/video/3"]
        "500":
          description: "Unexpected error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: { code: 500, message: "Something unexpected happened" }
    post:
      summary: "Add a new video file"
      operationId: postVideo
      requestBody:
        description: "The whole video in binary"
        content:
          multipart/form-data:
            schema:
              properties:
                video:
                  type: string
                  format: binary
      responses:
        "201":
          description: "Video uploaded"
          content:
            application/json:
              schema:
                type: string
              example: "/uploads/video/4"
        "500":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: { code: 500, message: "Something unexpected happened" }
  /api/files/gridfs/video/{id}:
    get:
      summary: "Get specific video in binary format from a bucket"
      operationId: getVideoFile
      parameters:
        - name: id
          in: path
          description: ID of the image
          required: true
          schema:
            type: integer
            format: int32
          example: 1
      responses:
        "200":
          description: "Video is fetched successfully"
    delete:
      summary: "Delete a video from the server"
      operationId: deleteVideo
      parameters:
        - name: id
          in: path
          description: ID of the video
          required: true
          schema:
            type: integer
            format: int32
          example: 1
      responses:
        "200":
          description: "Video successfully deleted"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              example: ["/uploads/video/2", "/uploads/video/3"]
        "404":
          description: "Video not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                {
                  code: 404,
                  message: "The video file you are trying to delete is not found"
                }
components:
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
security:
  - BasicAuth: []